fix type annotation wknml/nml_utils.py

+++ wknml/nml_utils.py
@@ -51,7 +51,7 @@ def detect_max_node_id_from_all_graphs(graph_dict: Dict[str, nx.Graph]) -> int:
     return max_id
 
 
-def approximate_minimal_edge_length(nml_or_graph: Union[NML, Tuple[Dict[str, List[nx.Graph]], Dict]], max_length: int, max_angle: float) -> Union[NML, Tuple[Dict[str, List[nx.Graph]], Dict]]:
+def approximate_minimal_edge_length(nml_or_graph: Union[NML, Tuple[Dict[str, List[nx.Graph]], Dict]], max_length: float, max_angle: float) -> Union[NML, Tuple[Dict[str, List[nx.Graph]], Dict]]:
     # it is easier to operate on a graph
     if isinstance(nml_or_graph, NML):
         nml_graph, parameter_dict = generate_graph(nml_or_graph)
